gestion_ventes/requirements.txt
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.3
django-environ==0.12.0
django-sendgrid-v5==1.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
pyjwt==2.9.0
coreapi==2.3.3
coreschema==0.0.4
uritemplate==4.1.1
gunicorn==23.0.0
pillow==11.0.0
psycopg==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sendgrid==6.11.0
sqlparse==0.5.2
urllib3==2.2.3
uvicorn==0.34.0
whitenoise==6.8.2
